:card_file_box: [BE] rename : 변수명 no를 id로 변경

+++ back/munggle/src/main/java/com/munggle/openAPI/repository/KindRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface KindRepository extends JpaRepository<Kind, Integer> {
+public interface KindRepository extends JpaRepository<Kind, Long> {
 
     List<Kind> findByKindNmLike(String input);
 }
