+++ back/munggle/src/main/java/com/munggle/openAPI/repository/LostDogRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LostDogRepository extends JpaRepository<LostDog, Integer> {
+public interface LostDogRepository extends JpaRepository<LostDog, Long> {
 
     List<LostDog> findByCareAddrLikeAndKindLike(String careAddr, String kind);
 }
